requirements.txt
Django==5.0.6
djangorestframework==3.15.2
mssql-django==1.5
pyodbc==5.2.0
python-dotenv==1.0.1
django-cors-headers==4.4.0
django-filter==24.2
drf-spectacular==0.27.2
